7419 statemachine

Máquina de estados para produção de conteúdos do TCC

See merge request !114

+++ db/migrate/20141028161522_add_grade_to_tccs.rb
@@ -1,6 +1,6 @@
 class AddGradeToTccs < ActiveRecord::Migration
   def change
-    add_column :tccs, :grade, :float
+    add_column :tccs, :grade, :integer
     add_column :tccs, :grade_updated_at, :datetime
 
     add_column :tcc_definitions, :moodle_instance_id, :integer
